Show Loading On Initial & Fix "Refresh" Button.
Previously the "Refresh" Button wouldn't capture touch events because it
was behind the RecyclerView. Before when it was above it, it made "Swipe
to Refresh" impossible as it caught the touch event instead.

This commit brings it to the front, but changes the size to Wrap_Content
so that it only captures the required touch events, and allows "Swipe to
Refresh" to continue working.

This commit also calls `AppbarRefreshLayout#setRefreshing(Boolean)` in
`showProgress(Integer)`, making the initial load display the
`SwipeRefreshLayout` loader.