Promote "open in browser" to default sign in method, remove "Basic Authentication"
Hello, I'm from the GitHub security team. We are in the process of removing the ability to use a password to authenticate to the GitHub API. We noticed this application supports and encourages in-app password-based authentication.

We suggest you only enable the "Login with your default browser (OAuth)" option. Let us know if we can help you in any way.

Ref: https://developer.github.com/changes/2020-02-14-deprecating-password-auth/
https://developer.github.com/changes/2020-02-14-deprecating-oauth-auth-endpoint/